+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Patch]** `ArrayType#read` now uses `itemType.read` (instead of `.readTrusted`).
+
 # 0.5.0-alpha.6
 
 - **[Feature]** Implement `MapType`